+++ gloptimize_vcf.h
@@ -8,7 +8,7 @@
 
 #include "genozip.h"
 
-extern const char *gl_optimize_dictionary (VBlockVCFP vb, BufferP dict, MtfNodeP nodes, unsigned dict_start_char, unsigned num_words);
+extern const char *gl_optimize_dictionary (VBlockVCFP vb, BufferP dict, MtfNodeP nodes, uint64_t dict_start_char, unsigned num_words);
 extern void gl_deoptimize_dictionary (char *data, int len);
 
 extern bool gl_optimize (const char *snip, unsigned len, char *updated_snip, unsigned *updated_len);
